Use fixed-width integer for UDP port.
Resolves Visual Studio warning about assigning an int to a CipUint.

+++ source/src/cip/cipioconnection.c
@@ -495,7 +495,7 @@ EipStatus OpenProducingMulticastConnection(
     return kEipStatusError;
   }
 
-  int port = htons(kOpenerEipIoUdpPort);
+  uint16_t port = htons(kOpenerEipIoUdpPort);
   if(kCipItemIdSocketAddressInfoTargetToOriginator != sock_addr_info->type_id) {
     port = sock_addr_info->sin_port;
   }
